requirements-on-server.txt
Package             Version
------------------- -----------
asgiref             3.8.1
boto3               1.38.1
botocore            1.38.1
cffi                1.17.1
cryptography        44.0.1
Django              5.1.6
django-cors-headers 4.7.0
djangorestframework 3.15.2
dnspython           2.7.0
gunicorn            21.2.0
jmespath            1.0.1
packaging           25.0
pip                 26.0.1
pycparser           2.22
PyJWT               2.10.1
pymongo             4.11.1
python-dateutil     2.9.0.post0
python-http-client  3.3.7
RapidFuzz           3.14.3
s3transfer          0.12.0
sendgrid            6.11.0
six                 1.17.0
sortedcontainers    2.4.0
sqlparse            0.5.3
starkbank-ecdsa     2.2.0
urllib3             2.4.0
